requirements.txt
numpy==1.19.5
scipy==1.5.4
scikit-learn==0.24.2
gensim==4.1.2
matplotlib==3.3.4
Pillow==8.4.0
pytesseract==0.3.8
reportlab==3.6.12
joblib==1.1.1
threadpoolctl==3.1.0
